feat: change precision

+++ src/vector.rs
@@ -2,19 +2,19 @@ use std::ops::{Neg, Add, Sub, Mul, Div};
 
 #[derive(Debug, Clone, Copy)]
 pub struct Vector {
-  pub x: f32,
-  pub y: f32,
-  pub z: f32,
+  pub x: f64,
+  pub y: f64,
+  pub z: f64,
 }
 
 impl Vector {
-  pub fn new(x: f32, y: f32, z: f32) -> Vector {
+  pub fn new(x: f64, y: f64, z: f64) -> Vector {
     Vector { x: x, y: y, z: z }
   }
 }
 
-pub trait Dot: Sized + Copy + Div<f32, Output = Self> {
-  fn dot(self, Self) -> f32;
+pub trait Dot: Sized + Copy + Div<f64, Output = Self> {
+  fn dot(self, Self) -> f64;
 }
 
 pub trait Cross {
@@ -22,8 +22,8 @@ pub trait Cross {
 }
 
 pub trait Norm {
-  fn norm(self) -> f32;
-  fn sqr_norm(self) -> f32;
+  fn norm(self) -> f64;
+  fn sqr_norm(self) -> f64;
 }
 
 pub trait Normalize {
@@ -33,17 +33,17 @@ pub trait Normalize {
 impl<T> Norm for T
   where T: Copy + Dot
 {
-  fn norm(self) -> f32 {
+  fn norm(self) -> f64 {
     self.sqr_norm().sqrt()
   }
 
-  fn sqr_norm(self) -> f32 {
+  fn sqr_norm(self) -> f64 {
     self.dot(self)
   }
 }
 
 impl<T> Normalize for T
-  where T: Copy + Norm + Div<f32, Output = Self>
+  where T: Copy + Norm + Div<f64, Output = Self>
 {
   fn normalize(self) -> T {
     self / self.norm()
@@ -51,7 +51,7 @@ impl<T> Normalize for T
 }
 
 impl Dot for Vector {
-  fn dot(self, rhs: Vector) -> f32 {
+  fn dot(self, rhs: Vector) -> f64 {
     self.x * rhs.x + self.y * rhs.y + self.z * rhs.z
   }
 }
@@ -90,15 +90,15 @@ impl Sub for Vector {
   }
 }
 
-impl Mul<f32> for Vector {
+impl Mul<f64> for Vector {
   type Output = Vector;
 
-  fn mul(self, rhs: f32) -> Vector {
+  fn mul(self, rhs: f64) -> Vector {
     Vector::new(self.x * rhs, self.y * rhs, self.z * rhs)
   }
 }
 
-impl Mul<Vector> for f32 {
+impl Mul<Vector> for f64 {
   type Output = Vector;
 
   fn mul(self, rhs: Vector) -> Vector {
@@ -114,15 +114,15 @@ impl Mul for Vector {
   }
 }
 
-impl Div<f32> for Vector {
+impl Div<f64> for Vector {
   type Output = Vector;
 
-  fn div(self, rhs: f32) -> Vector {
+  fn div(self, rhs: f64) -> Vector {
     Vector::new(self.x / rhs, self.y / rhs, self.z / rhs)
   }
 }
 
-impl Div<Vector> for f32 {
+impl Div<Vector> for f64 {
   type Output = Vector;
 
   fn div(self, rhs: Vector) -> Vector {
